Externalize some classes for custom Exceptions, custom types, and Enums to types.py
