+++ admin_cohort/types.py
@@ -1,0 +1,7 @@
+class UserInfo:
+    def __init__(self, username: str, email: str,
+                 firstname: str, lastname: str, **kwargs):
+        self.username = username
+        self.email = email
+        self.firstname = firstname
+        self.lastname = lastname
