+++ requirements.txt
@@ -3,7 +3,7 @@ djangorestframework==3.11.0
 markdown==3.1.1
 django-filter==2.2.0
 coreapi==2.3.3
-djangorestframework_simplejwt==4.4.0
+#djangorestframework_simplejwt==4.4.0
 django-rest-swagger==2.2.0
 drf-api-tracking==1.8.0
 drf-yasg==1.20.0
